wayland-protocols: Depend on wayland
Currently the lack of this expressed dependency does not cause
builds to break because we require wayland-egl as a sysdep, and
this will bring in a base wayland on the host system to satisfy
the dependency.

Since we do build both wayland and wayland-protocols, it's just
better to build wayland-protocols against the fresh wayland
that we build ourselves.