Allow running jhbuild as root
In some environments, like a Docker container, it's perfectly acceptable
to run Jhbuild as root; in some case, it's actually required in order to
work at all without hacking around the file permissions.

For this reason, we should modify the UID check performed at startup,
and if there is a specific variable in the environment, keep running
even as root.

https://bugzilla.gnome.org/show_bug.cgi?id=783965