apps: Remove Incubator
seahorse is beyond hope of repair, so it needs to be moved to world. It
will eventually be replaced with Credentials.

So Dictionary is the only app left in Incubator. It doesn't seem
worthwhile to keep this category around for just one app. And I'm not
really sure if Dictionary should ever really be in core anyway.
Incubator is just not useful anymore. Maybe it never was!