jhbuild/utils/systeminstall.py: added SystemInstall.install_options and implemented it it AptSystemInstall jhbuild/commands/sysdeps.py: added --install-options to CLI
Added SystemInstall.install_options and exposed it in the CLI

In some environments like headless CI services it's necessary to be able to
have more control over system package installation commands which are used in
implementations of SystemInstall, e.g. by adding --assume-yes to apt-get in
order to avoid it to wait for configuration decisions.

https://bugzilla.gnome.org/show_bug.cgi?id=788220