Add the ability to override tarball names
Modules hosted on GitHub have autogenerated tarball names coming from
the release tag. This means a lot of them have unhelpful names like
'v1.0.1.tar.gz', or '296.zip'.

If two different modules have the same tag, we end up with a collision
when trying to check that the tarball downloaded have the same checksum.

We can add a simple XML attribute that allows us to rename the tarball
if that's the case; we can also throw in variables support for the name
of the module and the version, which will hopefully help us to avoid
collisions.

https://bugzilla.gnome.org/show_bug.cgi?id=788779