core-3.28: gvfs now uses meson
Although autotools is still present, it will be removed on the
future. This patch changes the default build system to meson.

One of the changes of meson is that auto options have been changed
to boolean options, so unsupported features must be explicitly
disabled. In order to test all the available features in gvfs, this
patch also adds support for new system dependencies.

https://bugzilla.gnome.org/show_bug.cgi?id=789736