3.28: Enable wayland and udev condition on FreeBSD
FreeBSD has wayland in ports since a year ago although most packages are
still not built with wayland enabled by default. I think it is time to
test GNOME with wayland on FreeBSD.

FreeBSD also has a new port called 'libudev-devd', which is intended to
be a API-compatible libudev implementation on the top of FreeBSD devd.
Although there are only a few packages using it, it is already good
enough to allow building libgudev successfully without patches.

However, since downstream patches are still required to build wayland on
FreeBSD, we have to rely on the version installed in the system. To get
modulesets work with FreeBSD, all <dep package="wayland"/> are moved to
nested <if condition-set="linux"/> blocks because we are unable to build
wayland in JHBuild on FreeBSD.

libudev-devd on FreeBSD doesn't implement udev hwdb API. A new condition
'udev-hwdb' is added to Linux to indicate that its udev implementation
supports hwdb because there are a few modules depending on hwdb.

https://bugzilla.gnome.org/show_bug.cgi?id=792379