3.28: Rework conditions of dependencies and enable OpenGL for cairo
There were many dependencies put in the linux condition block, which was
wrong because these libraries are portable and not specific to features
in Linux kernel. X11 libraries are now put in the x11 condition block,
and things that are only useful on free operating systems are put in
the quartz unset condition block.

The use of quartz unset condition here is not precise. What we really
need is a way to enable them on free desktop environments, but at least
it works and it is what the old modulesets use.

WebKit needs the OpenGL backend of cairo. We have to enable it because
it is not enabled by default. I don't know whether the OpenGL backend
works on macOS. The only thing I know is that the bug report referenced
by the old modulesets is still not resolved.