3.28: Patch meson to fix library path issues on FreeBSD
All patches added here are already committed to the master branch of
meson and they will be available in the next release.

The first four patches are required for gtk-doc and g-ir-scanner to work
correctly on FreeBSD. These patches are not platform-specific, but it is
much likely to see build failure caused by wrong order of library paths
on FreeBSD than on GNU/Linux. Without them, many modules cannot be built
on FreeBSD, including essential ones such as ATK.

The last patch causes meson to show warnings when arguments starting
with '-Wl,' are used in cc.has_argument. I haven't seen any build
failure caused by this kind of wrong usage of cc.has_argument, but I
think it is better to let more people be aware of the common mistake.