Create a symlink to the srcdir from the builddir
When doing non-srcdir builds, the number one complaint is that it's hard
to go back to the source directory when inside a shell opened by jhbuild
on failure.

For this reason, if the build and source directories do not match,
jhbuild should create a symbolic link to the source directory inside the
build directory. This link exists while the user shell is active, and is
removed once the user exits; to avoid naming collisions, the symbolic
link is called `.jhbuild-srcdir`.