core-deps-latest: Use SpiderMonkey's bundled libicu
I believe that for SpiderMonkey 52 the version of libicu was already not
sufficient, as the --with-system-icu configure option wasn't given. This
removes that configure option as well as the dependency on libicu.

This commit is intended to be reverted when we can depend on a later
libicu in our build environment.