Remove circular dependency of gtk-doc on glib
Commit 59e38a6f678f8c238aa4d6ba652df1f8afc3fd57 added glib as a
dependency of gtk-doc, but glib depends on gtk-doc, which means we have
a circular dependency.

This is not a huge deal for Linux; gtk-doc is available on all
distributions, so it can be installed on the system. On non Linux
systems, though, this is a lot more tricky.

The reason why gtk-doc depends on glib is to run part of its test suite;
since this is not a requirement for jhbuild, we should disable the
gtk-doc tests, and remove the circular dependency.