configure.ac: Exit early if PYTHON is not Python 2
JHBuild still only supports Python 2. However, automake doesn't seem to
provide a way to choose between python 2 and 3, and it just accepts the
default one if its version number is high enough. Therefore, on systems
using Python 3 by default, JHBuild will be configured to use Python 3,
causing JHBuild to crash immediately because of invalid syntax. In order
to avoid installing a broken version of JHBuild, we should check the
version ourselves and exit early if PYTHON is not Python 2.