systeminstall: Support 'python3' sysdeps
Using code similar to 'python2' sysdeps, we add 'python3' sysdeps in
order to allow a module to depend on a system-installed Python 3 module.
Since JHBuild itself is written in Python 2, we cannot test Python 3
modules directly in the JHBuild process. Therefore, tests for Python 3
modules have to be done in a new process, and it is expected for python3
tests to be slower than python2 tests.