Remove buildbot support
Looking at the buildbot imports in that code it looks like it last
worked with buildbot 0.7.12, which was released 10 years ago and broke with 0.8.

This is also around the time it was last used in the gnome infra (build.gnome.org),
and the last buildbot related bug was filed in early 2011, so likely no one is using it since then.

We need to port to Python 3 and without any way to test this code
it will break anyway, so might as well remove it now.

If this for some reason breaks anything that has been working before, please file a bug!

See #26