Improve dynamic import error handling
Replace all code trying to import a module and skipping the import if
the module isn't there with try_import_module(). This allows us to only
skip the module-not-there case and raise with all other errors.

This is mainly helpful with Python 3 porting where invalid imports/syntax
make the imports fail but those errors getting ignored and things failing
later on without context.

This also removes all other __import__() calls with importlib.import_module()
because it's an easier API.