core-deps-latest: Update meson to 0.53.0
Meson 0.53 is now able to use PKG_CONFIG_PATH to sort library paths.
Therefore, the local patch included in JHBuild can be dropped.

https://github.com/mesonbuild/meson/issues/4271
https://github.com/mesonbuild/meson/pull/4325

Meson 0.53 changes the installation path of .pc files from lib/pkgconfig
to libdata/pkgconfig on FreeBSD without providing any way to override
the behavior. While it can be an improvement of usability on FreeBSD
because FreeBSD configures its pkgconf to use libdata/pkgconfig instead
of the upstream default lib/pkgconfig and share/pkgconfig, it can also
be a problem for JHBuild. JHBuild does not change the directory layout
depending on the platform, and doing so can confuse the user and make
the environment messy. Therefore, revert the change until there is a
better way to choose the default installation path.

https://github.com/mesonbuild/meson/pull/4410
https://github.com/mesonbuild/meson/pull/6343