base: also try for meson on autogenerated modules
When creating a non existant module from the default repository, don't
just try for autotools, but also for meson.

This is a copy-paste of the code for this same situation used in
`jhbuild make`.

This solves the annoyance of jhbuild'ing a GNOME module that is not in
the moduleset, but is on the default repo and already checked out. While
`jhbuild make` works fine, it requires you to be in the module's
directory, and the only workaround is to create your own custom
moduleset.