base: Search for 'meson.build' in the right path
Commit 2cc433db003c054b6a081047c48ef995f440d645 naively tested for the
cwd, assuming that `jhbuild buildone` would switch you to the checkout
path of the automatic module, so the condition was never met unless your
shell's cwd was the checked out module. Now it builds the expected path
of checkoutdir + modname.

This also inverts the if-else so that we search for autogen.sh instead
of meson.build. This means that when the code has not been checked out
yet (so that there's no files at all) this will default to meson.