mutter: require EGL/eglmesaext.h
Mutter started depending on EGL/eglmesaext.h, relevant
commit is:
https://gitlab.gnome.org/GNOME/mutter/-/commit/a444a4c5f58ea516ad3cd9d6ddc0056c3ca9bc90

But the jhbuild "egl" sysdep, which Mutter depends upon, is not
enough to check for that header (as it just checks for "egl.pc"
pkg-config module on the sytem).

So, we added a new sysdep "mesa-libEGL" that explicitly checks
for that header. I chose that name because the Fedora
package's name that provides that header is 'mesa-libEGL-devel'.

Issue #76 (the part of Mutter fail)