core-latest: Use main branch for NetworkManager
NM apparently switched to using "main" as its primary branch name
instead of "master", and it looks like git doesn't automatically follow
that change when jhbuild does "git pull", so it sticks to the outdated
"master" branch.

Fix it by manually specifying the revision to "main" in the moduleset to
make sure git switches the branch.