jhbuild/commands: Disable interspersed args
The option parser previously did not stop parsing options after the
first positional argument. This would result in option parsing errors
when trying to pass arguments to the executable passed to `jhbuild run`.
For example `jhbuild run --in-builddir=mutter test --version` results in
`jhbuild: error: no such option: --version`.

Disable this behavior with disable_interspersed_args(). This should be
fine to enable globally because according to the help description of all
commands the positional arguments are always last.