jhbuild/commands/make: Compare canonical paths
`jhbuild make` tries to figure out which module to build by looking at
the current working directory relative to the checkout directory. The
paths can point to the same directory but not be prefixes of each other
if there is a symlink involved. Fix the issue by canonicalizing both the
CWD and the checkout path.

This fixes `toolbox run jhbuild make` on Silverblue.