core-deps-latest: Don't build session manager for pipewire
Pipewire defaults to pipewire-media-session and will build that as a
subproject by default. That session manager has a few issues since these
days the default is wireplumber. Most people are not exactly interested
in building wireplumber when building pipewire though and there's a
separate jhbuild module for that anyway.

So let's disable building a session manager in pipewire and just use the
one provided by the system, which is wireplumber these days.