modules: Respect version on path systemdependencies
The documentation and in-tree modulesets do this:

```xml
<systemmodule id="foo">
  <branch repo="system" version="1.2.3"/>
  <systemdependencies>
    <dep type="path" name="foo"/>
  </systemdependencies>
</systemmodule>
```

However the specifying of `version` did nothing.

Now JHBuild parses the output of `foo --version`
to try and detect the actual version and compare
that against our required version.

Example output of `jhbuild sysdeps`:

```
System installed packages which are new enough:
    automake (installed=1.16.5)
...
Required packages:
  System installed packages which are too old:
    foo (required=1.2.3, installed=0.1.0)
...
```