Build poppler from git
Many issues in Evince are really bugs in Poppler,
so make it easier for people to build Poppler from
JHBuild, hopefully this will attract more contributors
or at least facilitate things for existant ones.

For licensing issues, Poppler consists of three repos,
one for encoding files (to properly render CJK and
Cyrillic), one to host the pdf files used for testing, and
lastly the source repo.