modtypes: Add pip type
Many python packages move to pyproject.toml and these projects can't be
installed with the distutil modtype. We can't use the `build` and `installer`
python modules because they also have a pyproject.toml (see MR #223).

This MR adds a new `pip` modtype. Use the tarball repository to get a sdist
from PyPi. The sdist is then built to a wheel using `pip wheel`. Build
dependencies are automatically installed to a temporal virtual environment. The
wheel is saved to the build folder. Then the modtype searches for a wheel in
the build folder and passes the path of the wheel to `pip install` for
installation. `pip install` doesn't install runtime dependencies (`--no-deps`).
Each runtime dependency has to be specified as a separate module.