Eval version placeholder consistently accross Repository subclasses
Some module variables like 'module name' or 'checkoutdir' allow to set a
'version' placeholder (${version}), that it's later replaced for version
number if set.

This logic is scattered among Repository subclasses, and even for some
cases, like 'meson' module, doesn't seem to be working.

It's better to have a single method in the Repository parent class
that applies this logic.